+++ src/main/scala/org/fs/mael/core/backend/ui/BackendConfigUi.scala
@@ -1,0 +1,7 @@
+package org.fs.mael.core.backend.ui
+
+import org.fs.mael.core.config.InMemoryConfigStore
+
+trait BackendConfigUi {
+  def get(): InMemoryConfigStore
+}