+++ src/main/scala/org/fs/mael/core/config/ConfigChangedEvent.scala
@@ -1,0 +1,3 @@
+package org.fs.mael.core.config
+
+case class ConfigChangedEvent[T](oldValue: T, newValue: T)