+++ src/main/scala/org/fs/mael/ui/config/EmptyPreferencePage.scala
@@ -1,0 +1,7 @@
+package org.fs.mael.ui.config
+
+import org.eclipse.jface.preference.FieldEditorPreferencePage
+
+class EmptyPreferencePage extends MFieldEditorPreferencePage(FieldEditorPreferencePage.FLAT) {
+  def createFieldEditors(): Unit = {}
+}