+++ src/test/scala/org/fs/mael/core/list/DownloadListManagerSpec.scala
@@ -10,6 +10,7 @@ import scala.io.Source
 import org.fs.mael.core.Status
 import org.fs.mael.core.checksum.Checksum
 import org.fs.mael.core.checksum.ChecksumType
+import org.fs.mael.core.config.InMemoryConfigStore
 import org.fs.mael.core.entry.DownloadEntry
 import org.fs.mael.core.event.Events._
 import org.fs.mael.core.utils.CoreUtils._
@@ -30,7 +31,7 @@ class DownloadListManagerSpec
 
     val backend = new StubBackend
     val entries = Seq(
-      backend.create(new URI("uri1"), new File("/a1"), Some("fn1"), None, "comment1", Some(backend.defaultCfg)),
+      backend.create(new URI("uri1"), new File("/a1"), Some("fn1"), None, "comment1", Some(defaultCfg)),
       backend.create(new URI("uri2"), new File("/a2"), None, Some(Checksum(ChecksumType.SHA1, "1abcde")), "comment2", None)
     )
     val serializer: DownloadListSerializer = new DownloadListSerializer {
@@ -89,7 +90,7 @@ class DownloadListManagerSpec
 
     val backend = new StubBackend
     val entries = IndexedSeq(
-      backend.create(new URI("uri1"), new File("/a1"), Some("fn1"), None, "comment1", Some(backend.defaultCfg)),
+      backend.create(new URI("uri1"), new File("/a1"), Some("fn1"), None, "comment1", Some(defaultCfg)),
       backend.create(new URI("uri2"), new File("/a2"), None, Some(Checksum(ChecksumType.SHA1, "1abcde")), "comment2", None)
     )
 
@@ -117,5 +118,11 @@ class DownloadListManagerSpec
     assert(eventMgr.events.size === 6)
     assert(eventMgr.events(4) === Removed(entries(0)))
     assert(eventMgr.events(5) === Removed(entries(1)))
+  }
+
+  private def defaultCfg: InMemoryConfigStore = {
+    val cfg = new InMemoryConfigStore
+    cfg.initDefault(StubBackend.StubSetting)
+    cfg
   }
 }