+++ src/test/scala/org/fs/mael/test/stub/AbstractSimpleBackend.scala
@@ -1,15 +1,17 @@
 package org.fs.mael.test.stub
 
 import org.eclipse.swt.widgets.TabFolder
-import org.fs.mael.core.backend.Backend
-import org.fs.mael.core.backend.BackendConfigUi
+import org.fs.mael.core.backend.AbstractBackend
 import org.fs.mael.core.backend.BackendDownloader
-import org.fs.mael.core.config.InMemoryConfigManager
+import org.fs.mael.core.backend.ui.BackendConfigUi
+import org.fs.mael.core.config.ConfigStore
+import org.fs.mael.core.config.InMemoryConfigStore
 import org.fs.mael.core.entry.DownloadEntry
 
 abstract class AbstractSimpleBackend(
-  override val id: String
-) extends Backend {
+  override val id:        String,
+  override val globalCfg: ConfigStore = new InMemoryConfigStore
+) extends AbstractBackend {
   override def downloader: BackendDownloader = new BackendDownloader(id) {
     override def eventMgr = ???
     override def transferMgr = ???
@@ -21,9 +23,9 @@ abstract class AbstractSimpleBackend(
 
   def downloadStopped(de: DownloadEntry): Unit = {}
 
-  override def defaultCfg = new InMemoryConfigManager
-
-  override def layoutConfig(cfgOption: Option[InMemoryConfigManager], tabFolder: TabFolder, isEditable: Boolean) = new BackendConfigUi {
-    override def get(): InMemoryConfigManager = cfgOption getOrElse defaultCfg
+  override def layoutConfig(cfgOption: Option[InMemoryConfigStore], tabFolder: TabFolder, isEditable: Boolean) = new BackendConfigUi {
+    override def get(): InMemoryConfigStore = cfgOption getOrElse defaultCfg
   }
+
+  override def pageDescriptors = Seq.empty
 }