+++ src/test/scala/org/fs/mael/test/stub/StubBackend.scala
@@ -1,6 +1,7 @@
 package org.fs.mael.test.stub
 
 import java.net.URI
+
 import org.fs.mael.core.config.ConfigSetting
 
 class StubBackend